+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
 # CHANGELOG
 
+## 
+- October 23 , 2022
+```
+- Switched Attendance Log from MongoDB to FireStore.
+- Attendance now asks for session.
+```
 
 ## Initial Release
 - August 01 , 2022